fetched and displayed data/moved Countdown.jsx file from /home to /UI
