+++ src/components/explore/ExploreItems.jsx
@@ -1,9 +1,25 @@
-import React from "react";
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
-import AuthorImage from "../../images/author_thumbnail.jpg";
-import nftImage from "../../images/nftImage.jpg";
+import { useEffect } from "react";
+import axios from "axios";
+import Countdown from "../UI/Countdown";
 
 const ExploreItems = () => {
+  const [exploreItemsData, setExploreItemsData] = useState([]);
+
+  useEffect(() => {
+    async function fetchData() {
+      const { data } = await axios.get(
+        "https://us-central1-nft-cloud-functions.cloudfunctions.net/explore"
+      );
+      setExploreItemsData(data);
+    }
+
+    fetchData();
+  }, []);
+
+  console.log(exploreItemsData);
+
   return (
     <>
       <div>
@@ -14,24 +30,25 @@ const ExploreItems = () => {
           <option value="likes_high_to_low">Most liked</option>
         </select>
       </div>
-      {new Array(8).fill(0).map((_, index) => (
+
+      {exploreItemsData.map((elem) => (
         <div
-          key={index}
+          key={elem.id}
           className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
           style={{ display: "block", backgroundSize: "cover" }}
         >
           <div className="nft__item">
             <div className="author_list_pp">
               <Link
-                to="/author"
+                to={`/authot/${elem.authorId}`}
                 data-bs-toggle="tooltip"
                 data-bs-placement="top"
               >
-                <img className="lazy" src={AuthorImage} alt="" />
+                <img className="lazy" src={elem.authorImage} alt="" />
                 <i className="fa fa-check"></i>
               </Link>
             </div>
-            <div className="de_countdown">5h 30m 32s</div>
+            <Countdown expiryDate={elem.expiryDate} />
 
             <div className="nft__item_wrap">
               <div className="nft__item_extra">
@@ -52,22 +69,23 @@ const ExploreItems = () => {
                 </div>
               </div>
               <Link to="/item-details">
-                <img src={nftImage} className="lazy nft__item_preview" alt="" />
+                <img src={elem.nftImage} className="lazy nft__item_preview" alt="" />
               </Link>
             </div>
             <div className="nft__item_info">
               <Link to="/item-details">
-                <h4>Pinky Ocean</h4>
+                <h4>{elem.title}</h4>
               </Link>
-              <div className="nft__item_price">1.74 ETH</div>
+              <div className="nft__item_price">{elem.likes} ETH</div>
               <div className="nft__item_like">
                 <i className="fa fa-heart"></i>
-                <span>69</span>
+                <span>{elem.likes}</span>
               </div>
             </div>
           </div>
         </div>
       ))}
+
       <div className="col-md-12 text-center">
         <Link to="" id="loadmore" className="btn-main lead">
           Load more
@@ -78,3 +96,58 @@ const ExploreItems = () => {
 };
 
 export default ExploreItems;
+
+// {new Array(8).fill(0).map((_, index) => (
+// <div
+//   key={index}
+//   className="d-item col-lg-3 col-md-6 col-sm-6 col-xs-12"
+//   style={{ display: "block", backgroundSize: "cover" }}
+// >
+//   <div className="nft__item">
+//     <div className="author_list_pp">
+//       <Link
+//         to="/author"
+//         data-bs-toggle="tooltip"
+//         data-bs-placement="top"
+//       >
+//         <img className="lazy" src={AuthorImage} alt="" />
+//         <i className="fa fa-check"></i>
+//       </Link>
+//     </div>
+//     <div className="de_countdown">5h 30m 32s</div>
+
+//     <div className="nft__item_wrap">
+//       <div className="nft__item_extra">
+//         <div className="nft__item_buttons">
+//           <button>Buy Now</button>
+//           <div className="nft__item_share">
+//             <h4>Share</h4>
+//             <a href="" target="_blank" rel="noreferrer">
+//               <i className="fa fa-facebook fa-lg"></i>
+//             </a>
+//             <a href="" target="_blank" rel="noreferrer">
+//               <i className="fa fa-twitter fa-lg"></i>
+//             </a>
+//             <a href="">
+//               <i className="fa fa-envelope fa-lg"></i>
+//             </a>
+//           </div>
+//         </div>
+//       </div>
+//       <Link to="/item-details">
+//         <img src={nftImage} className="lazy nft__item_preview" alt="" />
+//       </Link>
+//     </div>
+//     <div className="nft__item_info">
+//       <Link to="/item-details">
+//         <h4>Pinky Ocean</h4>
+//       </Link>
+//       <div className="nft__item_price">1.74 ETH</div>
+//       <div className="nft__item_like">
+//         <i className="fa fa-heart"></i>
+//         <span>69</span>
+//       </div>
+//     </div>
+//   </div>
+// </div>
+// ))}
