+++ src/components/home/NewItems.jsx
@@ -4,7 +4,7 @@ import axios from "axios";
 import OwlCarousel from "react-owl-carousel";
 import "owl.carousel/dist/assets/owl.carousel.css";
 import "owl.carousel/dist/assets/owl.theme.default.css";
-import Countdown from "./Countdown";
+import Countdown from "../UI/Countdown";
 
 const NewItems = () => {
   const [newItemsData, setNewItemsData] = useState([]);
